# doc/api.yml
openapi: 3.0.3
info:
  title: HEI Grade Flow API
  description: |
    API de gestion des notes, parcours (EL/TN), groupes, validations de crédits
    et relevés de notes pour HEI.

    ### Authentification et Sécurité :
    - **Gestion de compte :** Aucun auto-enregistrement public (`/register`). Les comptes sont créés exclusivement par un **ADMIN**.
    - **Connexion :** Tous les utilisateurs (STUDENT, TEACHER, ADMIN) s'authentifient via `POST /api/auth/login` pour obtenir un jeton JWT.

    ### Matrice de Sécurité & Droits d'accès (RBAC) :
    - **STUDENT** : Peut consulter exclusivement SES notes, SES inscriptions (groupe, semestre, parcours EL/TN) et SES relevés de notes.
    - **TEACHER** : Peut consulter et modifier TOUTES les notes des cours qui lui sont affectés, créer des évaluations et publier les notes.
    - **ADMIN** : Accès total sur tout le système (création utilisateurs/étudiants/enseignants, inscriptions, affectations, exports Excel des diplômés sur 3 ans, audits).

  version: 3.0.0
servers:
  - url: http://localhost:8080
    description: Serveur de Développement Local
security:
  - BearerAuth: []
...
paths:
...
  /api/promotions:
    get:
      summary: Lister toutes les promotions
      tags: [Academic]
      responses:
        '200':
          description: Liste des promotions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PromotionResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
...
  /api/grades/{gradeId}:
    get:
      summary: Récupérer une note par son ID
      tags: [Grades]
      parameters:
        - name: gradeId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Détail de la note
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GradeResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          description: Note non trouvée
    put:
      summary: Modifier une note (avec motif et historisation automatique)
      tags: [Grades]
      parameters:
        - name: gradeId
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GradeUpdateRequest'
      responses:
        '200':
          description: Note modifiée
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GradeResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
...
  /api/students/{std}/transcript/send-email:
    post:
      summary: Générer le relevé PDF, le déposer sur S3 et l'envoyer par email à l'étudiant
      description: Crée une entrée DOCUMENT_EXPORT (docType=TRANSCRIPT_PDF) et déclenche l'envoi asynchrone.
      tags: [Transcripts & Exports]
      parameters:
        - name: std
          in: path
          required: true
          schema: { type: string, example: "STD21001" }
        - name: semesterCode
          in: query
          schema: { type: string, enum: [S1, S2, S3, S4, S5, S6] }
      responses:
        '202':
          description: Génération et envoi initiés
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentExportResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          description: Étudiant non trouvé
  /api/document-exports:
    get:
      summary: Consulter l'historique des exports (relevés PDF et listes Excel)
      tags: [Transcripts & Exports]
      parameters:
        - name: studentId
          in: query
          schema: { type: string, format: uuid }
        - name: promotionId
          in: query
          schema: { type: string, format: uuid }
        - name: academicYearId
          in: query
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Historique des exports
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DocumentExportResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /api/document-exports/{id}:
    get:
      summary: Suivre le statut d'un export asynchrone (PENDING → SUCCESS/FAILED)
      tags: [Transcripts & Exports]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Statut courant de l'export
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentExportResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          description: Export non trouvé
  /api/document-exports/graduates:
    post:
      summary: Générer l'export Excel des étudiants diplômés d'une promotion sur 3 ans (ADMIN uniquement)
      tags: [Transcripts & Exports]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [promotionId]
              properties:
                promotionId:
                  type: string
                  format: uuid
      responses:
        '202':
          description: Génération de l'export Excel initiée avec succès
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentExportResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Saisissez le token JWT généré par l'endpoint /api/auth/login
  responses:
    UnauthorizedError:
      description: Non authentifié
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    ForbiddenError:
      description: Accès interdit
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        statusCode: { type: integer, example: 403 }
        message: { type: string }
        timestamp: { type: string, format: date-time }
...
    PromotionResponse:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        year: { type: integer }
...
    GradeUpdateRequest:
      type: object
      required: [score, modifiedByUserId, reason]
      properties:
        score: { type: number, format: double }
        modifiedByUserId: { type: string, format: uuid }
        reason: { type: string }
    GradeResponse:
      type: object
      properties:
        id: { type: string, format: uuid }
        studentId: { type: string, format: uuid }
        evaluationId: { type: string, format: uuid }
        score: { type: number }
        published:
          type: boolean
          description: false = brouillon (visible enseignant seulement), true = visible étudiant.
        publishedAt: { type: string, format: date-time, nullable: true }
...
    DocumentExportResponse:
      type: object
      properties:
        id: { type: string, format: uuid }
        studentId: { type: string, format: uuid, nullable: true }
        promotionId:
          type: string
          format: uuid
          nullable: true
          description: Renseigné pour les exports GRADUATES_EXCEL, absent pour un TRANSCRIPT_PDF.
        academicYearId: { type: string, format: uuid }
        docType: { type: string, enum: [TRANSCRIPT_PDF, GRADUATES_EXCEL] }
        s3Key: { type: string }
        status: { type: string, enum: [PENDING, SUCCESS, FAILED] }
        generatedAt: { type: string, format: date-time }
        sentAt: { type: string, format: date-time, nullable: true }